requirements.txt
allennlp
python-chess
